http-api/requirements.txt
aiohttp==1.0.3
asyncio-nats-client==0.3.2
